pyproject.toml
[tool.poetry]
name = "todolistproject"
version = "0.1.0"
description = ""
authors = ["mvladlena85 <mvladlena@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
